Refactor AppHost/apphost
