+++ src/InterviewCoach.AppHost.Core/AgentMode.cs
@@ -1,0 +1,7 @@
+public enum AgentMode
+{
+    Unknown,
+    Single,
+    LlmHandOff,
+    CopilotHandOff
+}
