+++ src/InterviewCoach.AppHost.Core/LlmProvider.cs
@@ -1,0 +1,7 @@
+public enum LlmProvider
+{
+    Unknown,
+    AzureOpenAI,
+    MicrosoftFoundry,
+    GitHubCopilot
+}
