+++ src/InterviewCoach.AppHost.Core/ResourceConstants.cs
