+++ src/InterviewCoach.AppHost/InterviewCoach.AppHost.csproj
@@ -5,16 +5,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Aspire.Hosting.Azure" />
-    <PackageReference Include="Aspire.Hosting.Azure.AppContainers" />
-    <PackageReference Include="Aspire.Hosting.Foundry" />
-    <PackageReference Include="Aspire.Hosting.OpenAI" />
-    <PackageReference Include="Azure.Provisioning.Storage" />
     <PackageReference Include="CommunityToolkit.Aspire.Hosting.SQLite" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\InterviewCoach.AppHost.Core\InterviewCoach.AppHost.Core.csproj" IsAspireProjectResource="false" />
+    <ProjectReference Include="..\InterviewCoach.AppHost.Core\InterviewCoach.AppHost.Core.csproj" />
   </ItemGroup>
 
   <ItemGroup>
